wip: refactor.
