+++ src/FeaturePoint.tsx
@@ -1,7 +1,13 @@
+import axios from "axios";
+import { useState, useEffect } from "react";
 import { Marker, Popup } from "react-leaflet";
 import type { IFeaturePoint } from "./Map";
+const loadFeatures = async (url: string) => {
+  const res = await axios.get<[IFeaturePoint]>(url);
+  return res.data;
+};
 
-const featurePoint = (props: { point: IFeaturePoint }) => {
+const FeaturePoint = (props: { point: IFeaturePoint }) => {
   return (
     <Marker position={[props.point.lat, props.point.lng]}>
       <Popup>
@@ -15,4 +21,14 @@ const featurePoint = (props: { point: IFeaturePoint }) => {
   );
 };
 
-export default featurePoint;
+const FeaturePoints = (url: string) => {
+  const [features, setFeatures] = useState<IFeaturePoint[]>([]);
+  useEffect(() => {
+    loadFeatures(url).then((data) => setFeatures(data));
+  });
+  return features.map((feature) => (
+    <FeaturePoint point={feature} key={feature.name} />
+  ));
+};
+
+export default FeaturePoints;
